roadmap_polygon.h:  use 32 bit ints for polygon counting
i inadvertently gave them 64 bits.  this breaks maps built previously
with the "lots and lots of polygons" fix.

+++ roadmap/src/roadmap_db_polygon.h
@@ -55,8 +55,8 @@ typedef struct {  /* table polygon/head */
 
 typedef struct {  /* table polygon/head */
 
-   unsigned long first;  /* 32 bits */
-   unsigned long count;  /* 32 bits */
+   unsigned int first;  /* 32 bits */
+   unsigned int count;  /* 32 bits */
 
    RoadMapString name;
    unsigned char cfcc;
